pyproject.toml
[tool.poetry]
name = "music-project"
version = "0.1.0"
description = ""
authors = ["1apostoli <79337131+1apostoli@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pymysql = "^1.1.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
